+++ src/ServiceBus/ServiceBus.Autorest/test/Set-AzServiceBusGeoDRConfigurationFailOver.Tests.ps1
@@ -15,13 +15,13 @@ if(($null -eq $TestName) -or ($TestName -contains 'Set-AzServiceBusGeoDRConfigur
 }
 
 Describe 'Set-AzServiceBusGeoDRConfigurationFailOver' {
-    It 'Fail'  {
+    It 'Fail' -skip:$($env.secondaryLocation -eq '')  {
         Set-AzServiceBusGeoDRConfigurationFailOver -ResourceGroupName $env.resourceGroup -NamespaceName $env.secondaryNamespace -Name $env.alias
-
+        
         while($drConfig.ProvisioningState -ne "Succeeded"){
             $drConfig = Get-AzServiceBusGeoDRConfiguration -Name $env.alias -ResourceGroupName $env.resourceGroup -NamespaceName $env.secondaryNamespace
             if ($TestMode -ne 'playback') {
-                Start-TestSleep 10
+                Start-Sleep 10
             }
         }
 
@@ -32,26 +32,26 @@ Describe 'Set-AzServiceBusGeoDRConfigurationFailOver' {
 
         $drConfig = Remove-AzServiceBusGeoDRConfiguration -Name $env.alias -ResourceGroupName $env.resourceGroup -NamespaceName $env.secondaryNamespace
         if ($TestMode -ne 'playback') {
-                Start-TestSleep 180
+                Start-Sleep 180
         }
         $drConfig = New-AzServiceBusGeoDRConfiguration -Name $env.alias -ResourceGroupName $env.resourceGroup -NamespaceName $env.primaryNamespace -PartnerNamespace $env.secondaryNamespaceResourceId
-
+        
         while($drConfig.ProvisioningState -ne "Succeeded"){
             $drConfig = Get-AzServiceBusGeoDRConfiguration -Name $env.alias -ResourceGroupName $env.resourceGroup -NamespaceName $env.primaryNamespace
             if ($TestMode -ne 'playback') {
-                Start-TestSleep 10
+                Start-Sleep 10
             }
         }
     }
-    It 'FailViaIdentity' {
+    It 'FailViaIdentity' -skip:$($env.secondaryLocation -eq '') {
         $drConfig = Get-AzServiceBusGeoDRConfiguration -Name $env.alias -ResourceGroupName $env.resourceGroup -NamespaceName $env.secondaryNamespace
-
+        
         Set-AzServiceBusGeoDRConfigurationFailOver -InputObject $drConfig
-
+        
         do {
             $drConfig = Get-AzServiceBusGeoDRConfiguration -Name $env.alias -ResourceGroupName $env.resourceGroup -NamespaceName $env.secondaryNamespace
             if ($TestMode -ne 'playback') {
-                Start-TestSleep 10
+                Start-Sleep 10
             }
         } while($drConfig.ProvisioningState -ne "Succeeded")
 
@@ -61,11 +61,11 @@ Describe 'Set-AzServiceBusGeoDRConfigurationFailOver' {
         $drConfig.Role | Should -Be "PrimaryNotReplicating"
 
         $drConfig = Remove-AzServiceBusGeoDRConfiguration -Name $env.alias -ResourceGroupName $env.resourceGroup -NamespaceName $env.secondaryNamespace
-
+        
         if ($TestMode -ne 'playback') {
-                Start-TestSleep 180
+                Start-Sleep 180
         }
 
-        { Get-AzServiceBusGeoDRConfiguration -Name $env.alias -ResourceGroupName $env.resourceGroup -NamespaceName $env.primaryNamespace } | Should -Throw
+        { Get-AzServiceBusGeoDRConfiguration -Name $env.alias -ResourceGroupName $env.resourceGroup -NamespaceName $env.primaryNamespace -ErrorAction Stop } | Should -Throw
     }
 }