+++ src/Sql/Sql/Database/Services/AzureSqlDatabaseAdapter.cs
@@ -26,6 +26,8 @@ using System.Linq;
 using Microsoft.Azure.Commands.Sql.Common;
 using Microsoft.Azure.Management.Sql.Models;
 using Microsoft.Rest.Azure.OData;
+using Microsoft.Azure.Management.Sql;
+using Newtonsoft.Json;
 
 namespace Microsoft.Azure.Commands.Sql.Database.Services
 {
@@ -196,7 +198,9 @@ namespace Microsoft.Azure.Commands.Sql.Database.Services
                 FederatedClientId = model.Database.FederatedClientId,
                 EncryptionProtectorAutoRotation = model.Database.EncryptionProtectorAutoRotation,
                 UseFreeLimit = model.Database.UseFreeLimit,
-                FreeLimitExhaustionBehavior = model.Database.FreeLimitExhaustionBehavior
+                FreeLimitExhaustionBehavior = model.Database.FreeLimitExhaustionBehavior,
+                ManualCutover = model.Database.ManualCutover,
+                PerformCutover = model.Database.PerformCutover
             });
 
             return CreateDatabaseModelFromResponse(resourceGroup, serverName, resp);
@@ -323,6 +327,7 @@ namespace Microsoft.Azure.Commands.Sql.Database.Services
                 var response = Communicator.ListOperations(resourceGroupName, serverName, databaseName);
                 IEnumerable<AzureSqlDatabaseActivityModel> list = response.Select((r) =>
                 {
+
                     return new AzureSqlDatabaseActivityModel()
                     {
                         DatabaseName = r.DatabaseName,
@@ -342,7 +347,8 @@ namespace Microsoft.Azure.Commands.Sql.Database.Services
                         },
                         EstimatedCompletionTime = r.EstimatedCompletionTime,
                         Description = r.Description,
-                        IsCancellable = r.IsCancellable
+                        IsCancellable = r.IsCancellable,
+                        OperationPhaseDetails = r.OperationPhaseDetails,
                     };
                 });