+++ src/Websites/Websites.Autorest/Properties/AssemblyInfo.cs
@@ -23,6 +23,6 @@ using System.Runtime.InteropServices;
 
 [assembly: ComVisible(false)]
 [assembly: CLSCompliant(false)]
-[assembly: Guid("6987ad96-09d8-4750-8333-c558508fa231")]
-[assembly: AssemblyVersion("3.2.0")]
-[assembly: AssemblyFileVersion("3.2.0")]
+[assembly: Guid("0725d62c-9555-4db5-b36c-b0156e4f88b8")]
+[assembly: AssemblyVersion("3.2.1")]
+[assembly: AssemblyFileVersion("3.2.1")]