+++ src/DnsResolver/DnsResolver/ChangeLog.md
@@ -18,7 +18,11 @@
         - Additional information about change #1
 -->
 ## Upcoming Release
+
+## Version 1.0.0
 * General availability for module Az.DnsResolver
+
+## Version 0.2.2
 * Introduced secrets detection feature to safeguard sensitive data.
 
 ## Version 0.2.1