+++ src/SelfHelp/SelfHelp/ChangeLog.md
@@ -18,6 +18,8 @@
         - Additional information about change #1
 -->
 ## Upcoming Release
+
+## Version 0.2.0
 * Added support for Discovery NLP API at Subscription and Tenant levels.
 * Added support for SelfHelp API and Discovery API at Tenant level.
 * Added support for Simplified Solutions API.